% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BernoulliNos.R
\name{BernoulliNos}
\alias{BernoulliNos}
\title{Compute Bernoulli numbers up to a given index.}
\usage{
BernoulliNos(n, lista_entera = TRUE)
}
\arguments{
\item{n}{Integer specifying the index up to which Bernoulli numbers should be computed.}

\item{lista_entera}{Logical indicating whether to return the sequence of Bernoulli numbers as a character vector (default: TRUE). If FALSE, only the last Bernoulli number in the sequence is returned as a character.}
}
\value{
If \code{lista_entera} is \code{TRUE}, returns a character vector containing the computed Bernoulli numbers up to index \code{n}. If \code{lista_entera} is \code{FALSE}, returns the last computed Bernoulli number as a character.
}
\description{
Compute Bernoulli numbers up to a given index using the recursive algorithm.
}
\examples{
BernoulliNos(25, F)
}
