% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bisection.R
\name{bisection}
\alias{bisection}
\title{Perform the Bisection Method to find a root of a given function within a specified interval.}
\usage{
bisection(expr, lowerLim, upperLim, iterations, digits = 15, stepByStep = F)
}
\arguments{
\item{expr}{An expression representing the function for which a root is to be found. The function should be defined in terms of `x`.}

\item{lowerLim}{The lower limit of the interval within which the root is sought.}

\item{upperLim}{The upper limit of the interval within which the root is sought.}

\item{iterations}{The maximum number of iterations to perform.}

\item{digits}{Number of significant digits to be displayed in the result (default: 15).}

\item{stepByStep}{Logical indicating whether to display intermediate results at each iteration (default: FALSE).}
}
\value{
If `stepByStep` is `TRUE`, prints a table of intermediate results and returns the approximate root.
        If `stepByStep` is `FALSE`, returns the approximate root directly.
}
\description{
This function applies the Bisection Method to find a root of a given function within a specified interval.
}
\examples{
bisection(expression(x^2-2), 0, 2, 6, 15, T)
}
