% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DividedDifferences.R
\name{divDiff}
\alias{divDiff}
\title{Calculate divided differences of a given dataset up to a specified order.}
\usage{
divDiff(A, order, showTable = F)
}
\arguments{
\item{A}{A matrix containing the dataset, where the first column represents the x-values and the second column represents the corresponding function values (f(x)).}

\item{order}{The order of divided differences to calculate.}

\item{showTable}{Logical indicating whether to display the divided differences table (default: FALSE).}
}
\value{
If `showTable` is `TRUE`, prints the divided differences table and returns the highest order divided difference (which corresponds to the Newton interpolating polynomial coefficient).
        If `showTable` is `FALSE`, returns the highest order divided difference directly.
}
\description{
This function calculates the divided differences of a given dataset up to a specified order using Newton's divided differences method.
}
\examples{
# Calculate and display the divided differences of a dataset up to order 6
x <- c(1, 2, 3, 4, 5, 6, 7)
fx <- c(144, 56, 35, 22, 78, 3, 17)
divDiff(cbind(x, fx), 6, showTable = TRUE)
}
