% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegSimp3.8.R
\name{IntegSimp3.8}
\alias{IntegSimp3.8}
\title{Perform numerical integration using Simpson's 3/8 rule.}
\usage{
IntegSimp3.8(fn, a, b, n)
}
\arguments{
\item{fn}{The function to be integrated.}

\item{a}{The lower limit of integration.}

\item{b}{The upper limit of integration.}

\item{n}{The number of subintervals (must be a multiple of 3).}
}
\value{
The approximate value of the integral of the function \code{fn} from \code{a} to \code{b} using Simpson's 3/8 rule.
}
\description{
This function performs numerical integration using Simpson's 3/8 rule to approximate the integral of a given function over a specified interval.
}
\examples{
# Define the function to be integrated (e.g., x^2 - 2)
fn <- function(x) {
  y <- x^2 - 2
  return(y)
}

# Perform numerical integration using Simpson's 3/8 rule with n = 12 subintervals
result <- IntegSimp3.8(fn, 0, 2, 12)
print(result)

}
