% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegSimpson.R
\name{IntegSimpson}
\alias{IntegSimpson}
\title{Simpson's Rule Numerical Integration}
\usage{
IntegSimpson(func, a, b, n)
}
\arguments{
\item{func}{The function to be integrated.}

\item{a}{The lower limit of integration.}

\item{b}{The upper limit of integration.}

\item{n}{The number of subintervals (must be even).}
}
\value{
The approximate value of the integral of the function \code{func} from \code{a} to \code{b} using Simpson's rule.
}
\description{
This function applies Simpson's rule to compute the integral of a function \code{func} from \code{a} to \code{b} with \code{n} subintervals.
Simpson's rule is a numerical method for approximating definite integrals using quadratic interpolating polynomials.
}
\examples{
# Define the function to be integrated (e.g., x^2 - 2)
func <- function(x) {
  y <- x^2 - 2
  return(y)
}

# Perform numerical integration using Simpson's rule with n = 12 subintervals
result <- IntegSimpson(func, 0, 2, 12)
print(result)
}
