% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interp.Algebraic.R
\name{interpAlgebraic}
\alias{interpAlgebraic}
\title{Algebraic Interpolation}
\usage{
interpAlgebraic(A, xs, showExpr = F, showMatrix = F, dec = 2, showGraph = F)
}
\arguments{
\item{A}{Matrix containing the data points (x, fx), where the first column represents x-values and the second column represents corresponding function values fx.}

\item{xs}{Values at which to evaluate the interpolated polynomial.}

\item{showExpr}{Logical indicating whether to display the interpolated polynomial expression (default: FALSE).}

\item{showMatrix}{Logical indicating whether to display the coefficients matrix (default: FALSE).}

\item{dec}{Number of decimal places to round coefficients when displaying the polynomial expression (default: 2).}

\item{showGraph}{Logical indicating whether to plot the interpolated polynomial (default: FALSE).}
}
\value{
A vector containing the interpolated function values at specified \code{xs}.
}
\description{
This function uses the Vandermonde matrix method to perform algebraic interpolation based on given data points (x, fx) in matrix \code{A}.
It computes the coefficients of the algebraic polynomial interpolant and evaluates the polynomial at specified values \code{xs}.
}
\examples{
# Define the data points (x, fx)
x <- c(1, 2, 3, 4, 5, 6, 7)
fx <- c(144, 56, 35, 22, 78, 3, 17)
A <- cbind(x, fx)

# Perform algebraic interpolation and evaluate at x = 5.5
interpAlgebraic(A, 5.5, showExpr = TRUE, showMatrix = TRUE, showGraph = TRUE)
}
