% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterpHermite.R
\name{interpHermite}
\alias{interpHermite}
\title{Hermite Polynomial Interpolation}
\usage{
interpHermite(
  A,
  xs,
  showTable = F,
  showExpr = F,
  showMatrix = F,
  showGraph = F
)
}
\arguments{
\item{A}{Matrix containing the data points (x, fx) and their derivatives (dx, dxx).}

\item{xs}{Values at which to evaluate the Hermite polynomial.}

\item{showTable}{Logical indicating whether to display the divided differences table (default: FALSE).}

\item{showExpr}{Logical indicating whether to display the interpolated Hermite polynomial expression (default: FALSE).}

\item{showMatrix}{Logical indicating whether to display the coefficient matrix (default: FALSE).}

\item{showGraph}{Logical indicating whether to plot the interpolated Hermite polynomial (default: FALSE).}
}
\value{
A vector containing the interpolated function values at specified \code{xs}.
}
\description{
This function computes the Hermite polynomial interpolation based on the data matrix \code{A}, which contains data points (x, fx) along with their derivatives (dx, dxx).
It evaluates the Hermite polynomial at the specified values \code{xs}.
}
\examples{
# Example 1: Hermite interpolation with given derivatives
x <- c(3, 2, 5, 6, 9)
fx <- c(27, 6, 135, 234, 783)
dx <- c(33, NA, 85, NA, NA)
dxx <- c(20, NA, NA, 15, NA)
A <- cbind(x, fx, dx, dxx)
interpHermite(A, 4.5, showTable = TRUE, showExpr = TRUE, showMatrix = TRUE, showGraph = TRUE)
# Example 2: Hermite interpolation with missing derivatives
x <- c(-1, 1, 2)
fx <- c(3, 3, 6)
dx <- c(1, NA, -1)
dxx <- c(NA, NA, 10)
A <- cbind(x, fx, dx, dxx)
interpHermite(A, 1.5, showTable = TRUE, showExpr = TRUE, showMatrix = TRUE, showGraph = TRUE)
}
