% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpLagrange.R
\name{interpLagrange}
\alias{interpLagrange}
\title{Lagrange Polynomial Interpolation}
\usage{
interpLagrange(A, xs, showExpr = F, showMatrix = F, showGraph = F)
}
\arguments{
\item{A}{Matrix containing the data points (x, fx).}

\item{xs}{Values at which to evaluate the Lagrange polynomial.}

\item{showExpr}{Logical indicating whether to display the interpolated Lagrange polynomial expression (default: FALSE).}

\item{showMatrix}{Logical indicating whether to display the coefficient matrix (default: FALSE).}

\item{showGraph}{Logical indicating whether to plot the interpolated Lagrange polynomial (default: FALSE).}
}
\value{
A vector containing the interpolated function values at specified \code{xs}.
}
\description{
This function computes the Lagrange polynomial interpolation based on the data matrix \code{A}, which contains data points (x, fx).
It evaluates the Lagrange polynomial at the specified values \code{xs}.
}
\examples{
# Example: Lagrange interpolation
x <- c(1, 2, 3, 4, 5, 6, 7)
fx <- c(144, 56, 35, 22, 78, 3, 17)
A <- cbind(x, fx)
interpLagrange(A, 5.5, showExpr = FALSE, showMatrix = FALSE, showGraph = FALSE)
}
