% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterpNewton.R
\name{interpNewton}
\alias{interpNewton}
\title{Newton Polynomial Interpolation}
\usage{
interpNewton(A, xs, showTable = F, showExpr = F, showMatrix = F, showGraph = F)
}
\arguments{
\item{A}{Matrix containing the data points (x, fx).}

\item{xs}{Values at which to evaluate the Newton polynomial.}

\item{showTable}{Logical indicating whether to display the divided differences table (default: FALSE).}

\item{showExpr}{Logical indicating whether to display the interpolated Newton polynomial expression (default: FALSE).}

\item{showMatrix}{Logical indicating whether to display the coefficient matrix (default: FALSE).}

\item{showGraph}{Logical indicating whether to plot the interpolated Newton polynomial (default: FALSE).}
}
\value{
A vector containing the interpolated function values at specified \code{xs}.
}
\description{
This function computes the Newton polynomial interpolation based on the data matrix \code{A}, which contains data points (x, fx).
It evaluates the Newton polynomial at the specified values \code{xs}.
}
\examples{
# Example: Newton interpolation
x <- c(1, 2, 3, 4, 5, 6, 7)
fx <- c(144, 56, 35, 22, 78, 3, 17)
A <- cbind(x, fx)
interpNewton(A, 5.5, showTable = FALSE, showExpr = FALSE, showMatrix = FALSE, showGraph = FALSE)
}
