% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegulaFalsi.R
\name{regulaFalsi}
\alias{regulaFalsi}
\title{Regula Falsi Root Finding}
\usage{
regulaFalsi(expr, lowerLim, upperLim, iterations, digits = 15, stepByStep = F)
}
\arguments{
\item{expr}{Expression representing the function for which the root is to be found.}

\item{lowerLim}{Lower limit of the interval for root searching.}

\item{upperLim}{Upper limit of the interval for root searching.}

\item{iterations}{Maximum number of iterations for the Regula Falsi method.}

\item{digits}{Number of significant digits to use in computations (default: 15).}

\item{stepByStep}{Logical indicating whether to display the iteration steps (default: FALSE).}
}
\value{
The estimated root of the function within the specified interval.
}
\description{
This function uses the Regula Falsi method to find a root of the function represented by the expression \code{expr} within the interval [\code{lowerLim}, \code{upperLim}].
It iteratively refines the estimate of the root based on linear interpolation until convergence or the specified number of iterations.
}
\details{
Perform the Regula Falsi method for root finding.

This function implements the Regula Falsi (False Position) method to find a root of the function represented by the expression \code{expr} within the interval [\code{lowerLim}, \code{upperLim}].
It iteratively refines the estimate of the root based on the linear interpolation between the function values at the interval endpoints.
}
\examples{
# Example: Regula Falsi method for root finding
regulaFalsi(expression(x^2 - 2), 0, 2, 6, digits = 15, stepByStep = TRUE)
}
