% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleDifferences.R
\name{simpleDiff}
\alias{simpleDiff}
\title{Simple Finite Differences}
\usage{
simpleDiff(A, order, showTable = F)
}
\arguments{
\item{A}{Matrix containing the function values. The first column should represent the x-values and the second column the corresponding function values.}

\item{order}{Order of the finite differences to be computed.}

\item{showTable}{Logical indicating whether to display the differences table (default: FALSE).}
}
\value{
The computed finite difference value of the specified order.
}
\description{
This function calculates the simple finite differences of a specified order for the function values represented in matrix \code{A}.
The order of differences determines how many times the differences are taken consecutively.
The result is returned as the value of the highest order difference.
}
\examples{
# Example: Compute simple finite differences
x <- c(1, 2, 3, 4, 5, 6, 7)
fx <- c(144, 56, 35, 22, 78, 3, 17)
A <- cbind(x, fx)
simpleDiff(A, 6, showTable = TRUE)
}
