% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplinesClamp.R
\name{SplinesClamp}
\alias{SplinesClamp}
\title{Cubic Spline with Specified Derivatives at Endpoints}
\usage{
SplinesClamp(
  A,
  deriv,
  xs,
  showMatrix = FALSE,
  showPoly = FALSE,
  showGraph = FALSE,
  enlarge = FALSE,
  limits.y = c(-100, 100)
)
}
\arguments{
\item{A}{Matrix containing the data points. The first column corresponds to the x-values and the second column to the f(x) values.}

\item{deriv}{Vector of length 2 specifying the derivatives at the endpoints. The first element corresponds to the derivative at the first endpoint,
and the second element corresponds to the derivative at the last endpoint.}

\item{xs}{The point at which the spline will be evaluated.}

\item{showMatrix}{Logical indicating whether to display the matrix used in spline computation (default: FALSE).}

\item{showPoly}{Logical indicating whether to display the polynomial coefficients (default: FALSE).}

\item{showGraph}{Logical indicating whether to display a graphical representation of the spline (default: FALSE).}

\item{enlarge}{Logical indicating whether to enlarge the graph window (default: FALSE).}

\item{limits.y}{Limits for the y-axis in the graphical representation (default: c(-100, 100)).}
}
\value{
The value of the cubic spline evaluated at \code{xs}.
}
\description{
This function computes the value of a cubic spline at a specified point \code{xs} using the given data points \code{A},
where the derivatives at the endpoints are specified. It provides options to display the matrix used in the spline computation,
the polynomial coefficients, and/or a graphical representation of the spline.
}
\examples{
# Example: Compute the cubic spline with specified derivatives at the endpoints at a point
x <- c(1, 2, 3, 4, 5, 6, 7)
fx <- c(144, 56, 35, 22, 78, 3, 17)
A <- cbind(x, fx)
deriv <- c(1, 0) # Derivative at first endpoint = 1, derivative at last endpoint = 0
SplinesClamp(A, deriv, 3.5)

}
