% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplinesNat.R
\name{SplinesNat}
\alias{SplinesNat}
\title{Natural Cubic Spline}
\usage{
SplinesNat(
  A,
  xs,
  showMatrix = FALSE,
  showPoly = FALSE,
  showGraph = FALSE,
  enlarge = FALSE,
  limits.y = c(-100, 100)
)
}
\arguments{
\item{A}{Matrix containing the data points. The first column corresponds to the x-values and the second column to the f(x) values.}

\item{xs}{The point at which the spline will be evaluated.}

\item{showMatrix}{Logical indicating whether to display the matrix used in spline computation (default: FALSE).}

\item{showPoly}{Logical indicating whether to display the polynomial coefficients (default: FALSE).}

\item{showGraph}{Logical indicating whether to display a graphical representation of the spline (default: FALSE).}

\item{enlarge}{Logical indicating whether to enlarge the graph window (default: FALSE).}

\item{limits.y}{Limits for the y-axis in the graphical representation (default: c(-100, 100)).}
}
\value{
The value of the natural cubic spline evaluated at \code{xs}.
}
\description{
This function computes the value of a natural cubic spline at a specified point \code{xs} using the given data points \code{A}.
Additionally, it provides options to display the matrix used in the spline computation, the polynomial coefficients,
and/or a graphical representation of the spline.
}
\examples{
# Example: Compute the natural cubic spline at a point and display the graph
x <- c(1, 2, 3, 4, 5, 6, 7)
fx <- c(144, 56, 35, 22, 78, 3, 17)
A <- cbind(x, fx)
SplinesNat(A, 3.5, showGraph = TRUE, enlarge = TRUE, limits.y = c(-50, 160))

}
